+++ packages/types/src/clusters/groups.d.ts
@@ -428,12 +428,12 @@ export declare namespace Groups {
     export const features: ClusterType.Features<Features>;
 
     /**
-     * @deprecated Use {@link Groups}.
+     * @deprecated Scheduled for removal in 0.19. Use {@link Groups}.
      */
     export const Cluster: ClusterType.WithCompat<typeof Groups, Groups>;
 
     /**
-     * @deprecated Use {@link Groups}.
+     * @deprecated Scheduled for removal in 0.19. Use {@link Groups}.
      */
     export const Complete: typeof Groups;
 
@@ -441,7 +441,7 @@ export declare namespace Groups {
 }
 
 /**
- * @deprecated Use {@link Groups}.
+ * @deprecated Scheduled for removal in 0.19. Use {@link Groups}.
  */
 export declare const GroupsCluster: typeof Groups;