+++ packages/types/src/tlv/TlvArray.ts
@@ -40,7 +40,7 @@ export class ArraySchema<T> extends TlvSchema<T[]> {
         super();
     }
 
-    /** @deprecated Part of old ClusterType() compat layer. */
+    /** @deprecated Part of old ClusterType() compat layer.  Scheduled for removal in 0.19. */
     override get element(): TlvSchema.Element {
         const result: TlvSchema.Element = { type: "list" };